+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 ### 0.1.1
 
 - Fixed Base64 encoded templates parsing error
+
+### 0.2.0
+
+- Added Lorem generating command
+- Added image generating command
+- Added avatar generating command
+- Added Chinese localization support